Remove call to get_relative_path_to_script() from init_db() and add data_path as argument instead
To allow more flexibility, e.g. when testing, reintroduce an argument to
indicate the path to where the data is found.
When the script is executed directly make this data_path the relative path
to the script extended by "/data". Otherwise, leave the option open to
use other relative paths.