Adapt Position's create_all_positions_fitting_data() to take command instead of kwargs
Instead of passing a kwargs dict to create_all_positions_fitting_data() now
a list of lists is passed. Each sublist is a command consisting of
argument name, comparator, and value.
This way also creation of position objects based on > or < are possible.

As a consequence of this change adapt Person's get_held_positions() to
call create_all_positions_fitting_data() with a list of lists instead of
a keyword-argument pair.