Revise volunteer_add.py and decouple it from the DB
Instead of calls to the DB directly, instances of Position and Volunteer
are used to add entries to the DB.
Split prompt_.*() functions that prompt for several inputs.
Add new prompt_.*() functions for new attributes.
Write add_new_volunteer() and add_new_position() which generate which prompt
for the data they need, generate an instance of Volunteer or Position,
respectively and subsequently add it to the db.
Update helpers.py's string_follows_input_pattern() to also check ids.