Write singleton DatabaseConnection class to be used as component for other classes connecting to the DB
Write DatabaseConnection as Singleton, as there should always only be one
open connection to the Database.
define static variable _instance to keep track of whether an instance already
exists.
Write _init() to obtain a connection and cursor.
Write close() to close the connection, remove references to the connection
and cursor and reset _instance to None.
Write commit() to commit any changes made to the DB.
Write execute to take in commands and execute them on the DB.