Simplify get_region_non_m_count()
Instead of discerning between ranges with one or two borders, look
up wether zip in question lies between or at the first and last element of the
range list.
If there is one element in the range, these coincide.
If there is two elements in the range, they define the range as before.