Add functionality to read out data received from db.mensa and print which volunteers moved
Extract first name, last name and zip code from the data table.
Use RegionFinder on the zip code to find the current region of the volunteer.
Use the name to create the current version of the volunteer from the DB.
If the volunteer exists, is active and has a different home region stored
then the one found on db.mensa.de print the name and the new region.